Moving ConvertError to its own file.
